refactor: :fire: Navbar arreglado y reportService eliminado
