+++ backend/src/reports/reports.module.ts
@@ -1,15 +1,12 @@
 import { Module } from '@nestjs/common';
 import { ReportsResolver } from './reports.resolver';
-import { ReportsService } from './reports.service';
 import { UsersModule } from 'src/users/users.module';
 import { MembershipsModule } from 'src/memberships/memberships.module';
 import { SuscriptionModule } from 'src/suscription/suscription.module';
 
 @Module({
-  providers: [ReportsResolver, ReportsService],
+  providers: [ ReportsResolver ],
   imports : [
-    UsersModule,
-    MembershipsModule,
     SuscriptionModule,
   ]
