+++ backend/src/reports/reports.resolver.ts
@@ -1,19 +1,17 @@
 import { Query, Resolver } from '@nestjs/graphql';
 import { Suscription } from 'src/suscription/entities/suscription.entity';
-import { ReportsService } from './reports.service';
 import { SuscriptionService } from '../suscription/suscription.service';
 
 @Resolver()
 export class ReportsResolver {
 
     constructor (
-        private readonly reportsService : ReportsService
-
+        private readonly suscriptionService : SuscriptionService
     ) {}
 
     @Query(() => [Suscription], { name : 'reportSuscription'})
     async reportSuscription() : Promise<Suscription[]>{
-        return this.reportsService.reportSuscriptions()
+        return this.suscriptionService.reportSuscriptions()
     }
 
 }
